+++ android/KMAPro/kMAPro/src/main/java/com/tavultesoft/kmapro/MainActivity.java
@@ -608,9 +608,6 @@ public class MainActivity extends AppCompatActivity implements OnKeyboardEventLi
         url = data.toString();
       }
       if (url != null) {
-        // URL contains KMP to download in background.
-        boolean isCustom = FileUtils.isCustomKeyboard(url);
-
         String filename = data.getQueryParameter(KMKeyboardDownloaderActivity.KMKey_Filename);
         if (filename == null) {
           FileUtils.getFilename(url);
@@ -1070,7 +1067,7 @@ public class MainActivity extends AppCompatActivity implements OnKeyboardEventLi
         KMManager.addKeyboard(this, keyboardInfo);
       }
     } else {
-      // Error notifications handled in LanguageListActivity
+      // Error notifications previously handled in LanguageListActivity
     }
   }