+++ common/predictive-text/worker/index.ts
@@ -30,7 +30,7 @@
  */
 
 /// <reference path="../message.d.ts" />
-/// <reference path="../../../web/source/text/kmwstring.ts" />
+/// <reference path="../node_modules/@keymanapp/keyboard-processor/src/text/kmwstring.ts" />
 /// <reference path="models/dummy-model.ts" />
 /// <reference path="word_breaking/ascii-word-breaker.ts" />
 /// <reference path="./model-compositor.ts" />
@@ -349,7 +349,7 @@ if (typeof module !== 'undefined' && typeof module.exports !== 'undefined') {
   module.exports['ModelCompositor'] = ModelCompositor;
 } else if (typeof self !== 'undefined' && 'postMessage' in self) {
   // Automatically install if we're in a Web Worker.
-  LMLayerWorker.install(self as DedicatedWorkerGlobalScope);
+  LMLayerWorker.install(self as any); // really, 'as typeof globalThis', but we're currently getting TS errors from use of that.
 } else {
   //@ts-ignore
   window.LMLayerWorker = LMLayerWorker;