+++ web/source/dom/touchAliasElement.ts
@@ -1,5 +1,3 @@
-// References the KMW string extensions.
-/// <reference path="../text/kmwstring.ts" />
 // References extra HTML definitions not included by default in TS.
 /// <reference path="../kmwexthtml.ts" />
 // References device-specific code checks (separable module from KMW)